requirements.txt
ollama==0.6.0
pymupdf4llm==0.0.27
chromadb==1.2.2
langchain_text_splitters==1.0.0
tiktoken==0.12.0
typer==0.20.0
python-dotenv==1.2.1